# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample of e-commerce API.",
        "title": "Picus Final Project",
        "contact": {
            "name": "Emre Çalışır"
        },
        "version": "1.0"
    },
    "paths": {
        "/basket": {
            "get": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Get basket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Basket"
                ],
                "summary": "Get basket",
                "responses": {
                    "200": {
                        "description": "basket is empty"
                    },
                    "401": {
                        "description": "You are not authorized for this action"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Update basket",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Basket"
                ],
                "summary": "Update basket",
                "parameters": [
                    {
                        "description": "Basket",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/basket.Basket"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item added successfully"
                    },
                    "401": {
                        "description": "You are not authorized for this action"
                    },
                    "422": {
                        "description": "Not enough stock from product"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Create an order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Basket"
                ],
                "summary": "Create an order",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "You are not authorized for this action"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/category": {
            "get": {
                "description": "Get all categories",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Get all categories",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/category.Category"
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Create categories from csv",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Category"
                ],
                "summary": "Create categories from csv",
                "parameters": [
                    {
                        "type": "file",
                        "description": "csv",
                        "name": "csv",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/order": {
            "get": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Get all orders",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get all orders",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/order.Order"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/order/{id}": {
            "put": {
                "security": [
                    {
                        "TokenJWT": []
                    }
                ],
                "description": "Cancel order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Cancel order",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order canceled"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "basket.Basket": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "product_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Product"
                    }
                },
                "total_price": {
                    "type": "number",
                    "example": 999.99
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/user.User"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "category.Category": {
            "type": "object",
            "required": [
                "type"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "product": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Product"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "test type"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "order.Order": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "is_canceled": {
                    "type": "boolean",
                    "example": false
                },
                "line_total": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        1.99,
                        2.99,
                        3.99
                    ]
                },
                "product_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1,
                        2,
                        3
                    ]
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/product.Product"
                    }
                },
                "total_price": {
                    "type": "number",
                    "example": 999.99
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/user.User"
                },
                "user_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "product.Product": {
            "type": "object",
            "required": [
                "name",
                "price",
                "stock",
                "type"
            ],
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string",
                    "example": "product name"
                },
                "price": {
                    "type": "number",
                    "example": 99.99
                },
                "stock": {
                    "type": "integer",
                    "example": 10
                },
                "type": {
                    "type": "string",
                    "example": "category"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "user.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Test Address"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string",
                    "example": "test@tst.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Test User"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone": {
                    "type": "string",
                    "example": "+05341234567"
                },
                "role": {
                    "type": "string",
                    "example": "admin"
                },
                "status": {
                    "type": "string",
                    "example": "active"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenJWT": {
            "type": "apiKey",
            "name": "TokenJWT",
            "in": "header"
        }
    }
}